#916 - runtime: integrate features
