+++ src/mu/features/core.rs
@@ -37,16 +37,16 @@ use {
 };
 
 lazy_static! {
-    pub static ref MU_SYMBOLS: RwLock<HashMap<String, Tag>> = RwLock::new(HashMap::new());
-    pub static ref MU_FUNCTIONS: Vec<CoreFunctionDef> = vec![
-        ("core-info", 0, Feature::mu_core_info),
-        ("process-fds", 0, Feature::mu_fds),
-        ("process-mem-res", 0, Feature::mu_mem_res),
-        ("process-mem-virt", 0, Feature::mu_mem_virt),
-        ("process-time", 0, Feature::mu_time),
-        ("time-units-per-second", 0, Feature::mu_time_units),
-        ("delay", 0, Feature::mu_delay),
-        ("ns-symbols", 1, Feature::mu_ns_symbols),
+    pub static ref CORE_SYMBOLS: RwLock<HashMap<String, Tag>> = RwLock::new(HashMap::new());
+    pub static ref CORE_FUNCTIONS: Vec<CoreFunctionDef> = vec![
+        ("core-info", 0, Feature::core_core_info),
+        ("process-fds", 0, Feature::core_fds),
+        ("process-mem-res", 0, Feature::core_mem_res),
+        ("process-mem-virt", 0, Feature::core_mem_virt),
+        ("process-time", 0, Feature::core_time),
+        ("time-units-per-second", 0, Feature::core_time_units),
+        ("delay", 0, Feature::core_delay),
+        ("ns-symbols", 1, Feature::core_ns_symbols),
     ];
 }
 
@@ -61,8 +61,8 @@ pub trait Core {
 impl Core for Feature {
     fn feature() -> Feature {
         Feature {
-            symbols: Some(&MU_SYMBOLS),
-            functions: Some(&MU_FUNCTIONS),
+            symbols: Some(&CORE_SYMBOLS),
+            functions: Some(&CORE_FUNCTIONS),
             namespace: "mu/core".into(),
         }
     }
@@ -99,18 +99,18 @@ impl Core for Feature {
 }
 
 pub trait CoreFunction {
-    fn mu_core_info(_: &Env, _: &mut Frame) -> exception::Result<()>;
-    fn mu_delay(_: &Env, _: &mut Frame) -> exception::Result<()>;
-    fn mu_fds(_: &Env, _: &mut Frame) -> exception::Result<()>;
-    fn mu_mem_res(_: &Env, _: &mut Frame) -> exception::Result<()>;
-    fn mu_mem_virt(_: &Env, _: &mut Frame) -> exception::Result<()>;
-    fn mu_time(_: &Env, _: &mut Frame) -> exception::Result<()>;
-    fn mu_time_units(_: &Env, _: &mut Frame) -> exception::Result<()>;
-    fn mu_ns_symbols(_: &Env, _: &mut Frame) -> exception::Result<()>;
+    fn core_core_info(_: &Env, _: &mut Frame) -> exception::Result<()>;
+    fn core_delay(_: &Env, _: &mut Frame) -> exception::Result<()>;
+    fn core_fds(_: &Env, _: &mut Frame) -> exception::Result<()>;
+    fn core_mem_res(_: &Env, _: &mut Frame) -> exception::Result<()>;
+    fn core_mem_virt(_: &Env, _: &mut Frame) -> exception::Result<()>;
+    fn core_time(_: &Env, _: &mut Frame) -> exception::Result<()>;
+    fn core_time_units(_: &Env, _: &mut Frame) -> exception::Result<()>;
+    fn core_ns_symbols(_: &Env, _: &mut Frame) -> exception::Result<()>;
 }
 
 impl CoreFunction for Feature {
-    fn mu_delay(env: &Env, fp: &mut Frame) -> exception::Result<()> {
+    fn core_delay(env: &Env, fp: &mut Frame) -> exception::Result<()> {
         env.argv_check("%core:delay", &[Type::Fixnum], fp)?;
 
         let delay = Fixnum::as_i64(fp.argv[0]);
@@ -129,7 +129,7 @@ impl CoreFunction for Feature {
         Ok(())
     }
 
-    fn mu_fds(env: &Env, fp: &mut Frame) -> exception::Result<()> {
+    fn core_fds(env: &Env, fp: &mut Frame) -> exception::Result<()> {
         let fds = fd_count_cur().unwrap();
 
         fp.value = Fixnum::with_u64(env, fds as u64)?;
@@ -137,7 +137,7 @@ impl CoreFunction for Feature {
         Ok(())
     }
 
-    fn mu_time(env: &Env, fp: &mut Frame) -> exception::Result<()> {
+    fn core_time(env: &Env, fp: &mut Frame) -> exception::Result<()> {
         fp.value = match cpu_time() {
             Ok(duration) => Fixnum::with_u64(env, duration.as_micros() as u64)?, // this is a u128
             Err(_) => panic!(),
@@ -146,13 +146,13 @@ impl CoreFunction for Feature {
         Ok(())
     }
 
-    fn mu_time_units(env: &Env, fp: &mut Frame) -> exception::Result<()> {
+    fn core_time_units(env: &Env, fp: &mut Frame) -> exception::Result<()> {
         fp.value = Fixnum::with_u64(env, 1000)?;
 
         Ok(())
     }
 
-    fn mu_mem_res(env: &Env, fp: &mut Frame) -> exception::Result<()> {
+    fn core_mem_res(env: &Env, fp: &mut Frame) -> exception::Result<()> {
         let vmem_info = get_process_memory_info().unwrap().resident_set_size;
 
         fp.value = Fixnum::with_u64(env, vmem_info * 4)?;
@@ -160,7 +160,7 @@ impl CoreFunction for Feature {
         Ok(())
     }
 
-    fn mu_mem_virt(env: &Env, fp: &mut Frame) -> exception::Result<()> {
+    fn core_mem_virt(env: &Env, fp: &mut Frame) -> exception::Result<()> {
         let vmem_info = get_process_memory_info().unwrap().virtual_memory_size;
 
         fp.value = Fixnum::with_u64(env, vmem_info * 4)?;
@@ -168,7 +168,7 @@ impl CoreFunction for Feature {
         Ok(())
     }
 
-    fn mu_core_info(env: &Env, fp: &mut Frame) -> exception::Result<()> {
+    fn core_core_info(env: &Env, fp: &mut Frame) -> exception::Result<()> {
         let version = env!("CARGO_PKG_VERSION");
         let alist = vec![
             Cons::cons(
@@ -194,7 +194,7 @@ impl CoreFunction for Feature {
         Ok(())
     }
 
-    fn mu_ns_symbols(env: &Env, fp: &mut Frame) -> exception::Result<()> {
+    fn core_ns_symbols(env: &Env, fp: &mut Frame) -> exception::Result<()> {
         let mut ns = fp.argv[0];
 
         if Tag::null_(&ns) {
