+++ src/mu/features/mod.rs
@@ -1,19 +1,13 @@
 //  SPDX-FileCopyrightText: Copyright 2022 James M. Putnam (putnamjm.design@gmail.com)
 //  SPDX-License-Identifier: MIT
 
-//! features module
+//! features crate
 #[cfg(feature = "core")]
 pub mod core;
 #[cfg(feature = "env")]
 pub mod env;
 pub mod feature;
-#[cfg(feature = "ffi")]
-pub mod ffi;
-#[cfg(feature = "nix")]
-pub mod nix;
 #[cfg(feature = "prof")]
 pub mod prof;
-#[cfg(feature = "std")]
-pub mod std;
-#[cfg(all(feature = "sysinfo", not(target_os = "macos")))]
-pub mod sysinfo;
+#[cfg(feature = "system")]
+pub mod system;
